kleaf: Fix initramfs modules lists test
The initramfs modules lists test is currently not comparing
the lists for equality; it is merely doing a check for None == None,
which is vacuously true. This is because the assertion compares
the return values of the sort() method on two lists.

The sort() method does not return anything, and instead sorts the
lists in place, so sort the lists first before doing a comparison.
Also, whitespace issues when reading the lists.

Fixes: Fixes: d08605f4c795 ("kleaf: Add support for specifying modules for recovery/charger")
Bug: 266752750
Change-Id: I6582bcc04efa5542725ca9be4df420b12919ebe0
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>