kleaf: Test modules.dep to ensure modules in modules.load* can be loaded
Depmod generates the modules.dep file which contains the dependencies
for each module in the directory that depmod runs in. When specifying
module lists, it is important that there's an entry for each module
in each module list to ensure that the module's dependencies are loaded
before attempting to load the module.

Therefore, add a test to ensure that each module in each module list
has an entry in modules.dep.

Bug: 323710246
Change-Id: Ie6a7cb65cad336d4ccda9fa6957624448690e3c7
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>