kleaf: Set rules_rust with dev_dependency = True.
If RUSTC_VERSION is set in kernel_toolchain_ext, kernel_build depends
on rules_rust, so rules_rust is needed for at least GKI.
However, if @kleaf is used as a dependent module in a DDKv2 set up and
drivers are building against prebuilts, kernel_build might not be
involved, and rules_rust might not be needed. If a DDKv2 driver is
building against a core kernel built from source, it needs to have
a direct dependency against rules_rust and call rust.toolchain() as well.

Bug: 326373147
Change-Id: I42c5c806a4980be4c84694e66e45f3ab64e4d1da