kleaf: Add additional_outs attr in kernel_modules_install
kernel_modules_install performs depmod on all modules.
However, since the result of the depmod can't be obtained,
an additional_outs attribute was created and changed
so that it can be obtained using copy_to_dist_dir.

It can be used as below:

kernel_modules_install(
    name = "foo_modules_install",
    kernel_modules = [":module_list"],
    additional_outs = ["modules.dep", "modules.alias"],
)

copy_to_dist_dir(
    name = "foo_dist",
    data = [
        ":foo_modules_install",
    ],
)

Test: bazel build //common:kernel_aarch64 --config=bzlmod
Bug: 327533166
Change-Id: Ia9efb26183553f72b5c27b8b0e1564702934b86a
Signed-off-by: Jaehyoung Choi <jkkkkk.choi@samsung.com>