kleaf: Mark _downloaded / _download_or_build targets deprecated.
Redirect users to use the single target under @gki_prebuilts.

This breaks the dependency of common_kernels on kernel_prebuilt_utils,
so there is one less thing to clean up.

These targets are also un-usable unless the @kleaf
module specifies the following in the Module.bazel
file:

   use_repo(kernel_prebuilt_ext, "gki_prebuilts")

However, with bzlmod, the repository `gki_prebuilts` may
or may not be defined depending on whether
kernel_prebuilt_ext.declare_kernel_prebuilts are called or not.

Test: TH
Bug: 291918087
Change-Id: I239cf5f2234f2c74b4a0b2b9146ecbc985272eee