kleaf: kernel_build's KernelEnvAndOutputsInfo does not include modules and implicit_outs.
As stated in the comments previously in the code, modules
and implicit_outs are not needed in KernelEnvAndOutputsInfo.
Hence drop them.

This refactors the code to use the common
get_env_and_outputs_info_setup_restore_outputs_command
function that restores outputs from ruledir to OUT_DIR.

This function can later be reused when implementing
mod_inst_env for kernel_filegroup.

Test: TH
Bug: 331700372
Change-Id: I6d12ef76d383c0eeb4b8654a6514df0a5822a8db