Revert "kleaf: Update docs. (kernel_filegroup.kernel_release)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7776f4a8-ff53-4f3f-9d9b-7d0c2bd5cf27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I074776a26963a9d5bce5e87b7156028ee9783a24