kleaf: cc_testing: delete full_static_link binaries.
The feature name is fully_static_link, not full_static_link.
However, the clang toolchain is not set up properly that
fully_static_link works, because the libunwind dependency
is not properly handled here. With fully_static_link, this
gives:

  ld.lld: error: undefined symbol: _Unwind_SetGR

Until we have a need for fully_static_link on this branch,
Delete the targets to avoid confusion.

fully_static_link is supported on android15-6.6 and above.

Test: build then readelf
Bug: 325514706
Change-Id: I3698f4efcf334b293c2c3d0f4642a58e493e2bb3