kleaf: Delete //build/kernel:hermetic-tools/*
The alises are turned into fail_rule() for
a sufficient amount of time. This change deletes
these fail_rule() targets. After this change,
command like the following:

bazel build //build/kernel:hermetic-tools/cp

will throw an error that says the target is
not found, instead of an error to use
hermetic_toolchain.

Bug: 291816237
Change-Id: I509a9e25af1f4bf6468155af565c3f30309b6e49