kleaf: Wrap hermetic tools with C++ binary.
Instead of symlinking to a script that uses
${0%/*}, use a native binary that executes

  $(readlink /proc/self/exe) $@ <hermetic_tools.extra_args>

This way, the tool can be executed anywhere
even when $0 is clobbered, e.g. when another
symlink is created against the tool. The only
remaining requirement is that this other symlink
must have the same tool_name.

In particular, this allows using
PATH={hermetic_tools} to run integration tests.

Bug: 338263410
Change-Id: If0ec635ef0c0bc7cfa4cf3758a5d6a8eedd8f556