kleaf: --repo_manifest supports repo != workspace root.
Change f63e8b029ee38458
("kleaf: fix case when workspace root is not the same as repo root.")
fixes the case that, when tools/bazel is invoked
without --repo_manifest, and repo root is not
the same as workspace root, --config=stamp still
properly infers the git metadata. This was not
the case with --repo_manifest and it was expected
to preprocess the repo manifest before feeding
into tools/bazel.

This change introduces a new format for
--repo_manifest:

  --repo_manifest=<REPO_ROOT>:<REPO_MANIFEST_PATH>

If this format is given, the repo root is
taken into consideration when parsing the repo
manifest so that --config=stamp still works.

This deprecates the invocation of `repo list -f`
and merge two code paths (with and without
--repo_manifest) into one that parses the manifest.

Add integration test for the issue by mounting
the workspace root in a different path, then
calling tools/bazel with a custom
--repo_manifest.

Bug: 281675905
Change-Id: Ib4f3e248dfb202aae46545c2ce16981e8bf50de1