kleaf: add --extra_git_projects
This adds extra projects on top of --repo_manifest
or `repo manifest` so that users can specify ones
not under `repo` control. In particular, if the
user is building a kernel from a KERNEL_DIR that
is a symlink, the stable-status.txt won't contain
a key with KERNEL_DIR. --extra_git_projects
allows users to inject custom keys.

Bug: 341223559
Test: integration_test
Change-Id: I5423608349faacfdadb4486110b0bb3c72ace736