kleaf: clean up abs paths in .config. (android14)
With a patch that allows relative paths of
UNUSED_KSYMS_WHITELIST to be interpreted against
the output directory, this change is possible to
remove absolute paths from .config, thereby removing
ir-reproducibility.

UNUSED_KSYMS_WHITELIST is stored in $OUT_DIR
before kernel_config is archived by
kernel_filegroup_declaration().

MODULE_SIG_KEY and SYSTEM_TRUSTED_KEY are keys, so
they are not distributed; they are restored when
re-setting up the environment for kernel_config,
and skipped for kernel_filegroup.

Test: TH
Bug: 333769605
Change-Id: I289563675e3d2ec182cf88801ee831ceda2ba3a9