kleaf: Option to clean up abs paths in .config.
With a patch that allows relative paths of
UNUSED_KSYMS_WHITELIST to be interpreted against
the output directory, this change is possible to
remove absolute paths from .config, thereby removing
ir-reproducibility.

UNUSED_KSYMS_WHITELIST is stored in $OUT_DIR
before kernel_config is archived by
kernel_filegroup_declaration().

MODULE_SIG_KEY and SYSTEM_TRUSTED_KEY are keys, so
they are not distributed; they are restored when
re-setting up the environment for kernel_config,
and skipped for kernel_filegroup.

Because the patch is not universally applied on
the frozen braches android14-5.15 and android14-6.1,
this is made an attribute,
kernel_build.allow_absolute_paths_in_config,
so non-frozen branches with the patch can opt in.

Test: TH
Bug: 333769605
Bug: 342390208
Change-Id: Ic2b65b3e6867a661fa306b19125787a98badf228