Revert "Kleaf: experimentally enable pigz as drop-in-replacement..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd20770d-ff00-493b-8e42-1a83ddc2aa5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2c00c9fc8865c12bb5d45348109529a1fa0d6a4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com