Revert "Add pigz repository including minimal build description"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd20770d-ff00-493b-8e42-1a83ddc2aa5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a3ea7243ab11818cb50616bf64a51a90b6ddee3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com