Revert "Add zopfli repository including minimal build description"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd20770d-ff00-493b-8e42-1a83ddc2aa5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I273d027316f51982e6487d3cb8a81ae8a8a8401e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com