kleaf: --incompatible_hermetic_actions
If the flag is set, attempt to enforce stronger
hermeticity for actions that don't specify
hermetic toolchain.

So far, the following list is added:
cp: needed for copy_file rule
python3: Needed to run any Python binary
  https://github.com/bazelbuild/bazel/issues/19355
uname: needed for rules_python
readlink/git/repo: needed for workspace_status.
  (git and repo is from host.)

Care must be taken to add more tools to this list.
This should only be reserved for actions
that are agnostic to Kleaf.

Bug: 348003050
Change-Id: Idc05f312ff697db7a0477a3a89b82813e0468653