Revert "Kleaf: ACK: Include download_configs file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7ba3f96-2edc-4cf0-bd9c-375262ed53a6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5a01e4e490d0a77ee5a477b188b74ee3d5a1f57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com