Revert "kleaf: init_ddk: sync git projects"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35c842a3-d11b-40db-a206-44f89f35fb1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9ac948483972121724cdf5908d6f755f04f0b03
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com